Fix compatibility with the latest versions of setuptools
The first change is to call fetch_build_eggs explicitly as in recent versions of setuptools the setup_requires does not appear to be respected. The second change is to explicitly parse the setup.cfg file as this appears to not be done inside Distribution anymore but at a higher level.