Only link object files in add_openmp_flags_if_available
Restrict the OpenMP test to attempting to link files that
end in the compiler's object file extension as determined
by `ccompiler.obj_extension`.

The OpenMP test was yielding a false negative if the user had
set the environment variable `CFLAGS='-coverage'` to measure
code coverage with GCC. The error message was:

    /usr/bin/ld:objects/test_openmp.gcno: file format not
    recognized; treating as linker script
    /usr/bin/ld:objects/test_openmp.gcno:1: syntax error

When GCC is called with the `-coverage` option, it emits extra
data files, which should *not* be passed to the linker because
they are not object files.