Extend exception handling of OpenMP support check
``check_openmp_support()`` compiles and links the small C code to test
OpenMP support. If either of these stages fail, their exceptions are caught and
``is_openmp_supported = False`` is set. However, they may still pass but the
code fails to run. ``CalledProcessError`` also needs to be caught to deal with
this.

Signed-off-by: James Noss <jnoss@stsci.edu>